--- a/pyproject.toml
+++ b/pyproject.toml
@@ -35,7 +35,8 @@ dependencies = [
     "numpy",
     "scipy",
     "pillow",
-    "scikit-image"
+    "scikit-image",
+    "aiofiles"
 ]
 
 [project.optional-dependencies]