This adds an `asyncio` mode for saving images (experimental). 

### Approach

- using a `asyncio.Queue` object, it appends all information to save each 2D slice to this queue
- it runs a co-routine which sets up 40 workers (`asyncio` `Task` objects - they are not actually threads)
- each worker takes out the information from the queue, converts the image into a binary buffer, and then writes it to disk asynchronously
- it uses `aiofiles` for allows async file writes (it is implemented using a thread pool behind the scenes)
- the main loop waits for all workers to finish and gather the exceptions
- The  co-routine is then executed using an event loop until it's done.

### Results

The results depend on disk used - it appears on a local disk, effects like faster disk writes and a RAM cache that is usually managed by the operating system, the existing method is faster. This was collected on Diamond, saving 1800 160x160 images in tiff format:

Preliminary results regarding async save_to_images  saving 1800 images of size 160x160 :
- On workstation, saving to `/tmp`:
    - old save_to_images, synchronous:  `762ms`
    - new with asyncio: `2,119ms`
    - ==> here it's much longer - the `/tmp` local disk is not a RAM-disk, but there are disk caches by the OS
- On the cluster node, saving to `/tmp`:
    - old version: `702ms`
    - asyncio version: `589ms`
- On the cluster node, saving to `/mnt/gpfs03/scratch/...`:
    - old version: `9,860ms`
    - new version: `741ms`

### Conclusion

- milage varies and depends on which type of application we optimise for primarily
- we also experimented with directly running all tasks (avoid the `Queue`), but that makes performance slightly worse
- the old version is still in the code in this PR, to enable a performance test and compare -  this is not intended as the final version