@yousefmoazzam 2024-03-15
Two notes regarding performance.

## Quick, and sometimes very quick, for smaller images

Mileage can apparently vary a lot in the case of running on a cluster node and saving to a GPFS mounted directory. The following `AssertionError`s are from running the performance test multiple times on data with shape `(160, 1800, 160)` and saving images along axis 1. The LHS values are the synchronous version, the RHS values are the asynchronous version. The times are in milliseconds.
```
AssertionError: assert 8810.761881999999 == 3215.963013
AssertionError: assert 7537.801946 == 3380.927038
AssertionError: assert 7543.671292 == 6608.081636999999
AssertionError: assert 7528.001184 == 3434.196302
AssertionError: assert 7760.500373999999 == 3573.0845759999997
AssertionError: assert 7517.687599999999 == 3828.923848
```

The following is another 6 runs of the performance test (possibly on a different cluster node, but to the same GPFS mounted directory) later the same day:
```
AssertionError: assert 6752.231022 == 1240.1117359999998
AssertionError: assert 6292.603373999999 == 1208.011518
AssertionError: assert 9494.373246 == 1204.732327
AssertionError: assert 9059.988963 == 1217.3358819999999
AssertionError: assert 7635.271691 == 1200.7767199999998
AssertionError: assert 8530.661401 == 1328.259562
```

Notice how the second set of runs has significantly quicker times for the asynchronous version, almost a third of the time from the first set of runs.

## Possible negligible difference when images get larger

When the images are larger, the difference in performance between the synchronous and asynchronous versions appears to be negligible. The following is from multiple runs of the performance test, but with the input data shape changed to `(2160, 1800, 2160)`:
```
AssertionError: assert 43847.911965 == 41457.283139
AssertionError: assert 40694.391481 == 43499.024356999995
AssertionError: assert 44484.593112999995 == 43720.287323
AssertionError: assert 46297.986830999995 == 46121.991662
AssertionError: assert 45993.563295 == 44426.137903999996
AssertionError: assert 47153.481657 == 43574.825004
```

Notice how there is not much difference between the two versions in this case.

Unlike the previous section where the times for the asynchronous version could vary drastically, these times consistently stay in the region of ~42 seconds across multiple sets of runs of the performance test.